Cleanup sharing leftovers
