+++ apps/files_sharing/src/additionalScripts.js
@@ -1,8 +1,6 @@
 import './share'
-import './sharetabview'
 import './sharebreadcrumbview'
 
-import './style/sharetabview.scss'
 import './style/sharebreadcrumb.scss'
 
 import './collaborationresourceshandler.js'
