--- core/js/merged-share-backend.js
@@ -1,9 +1,0 @@
-import './shareconfigmodel.js'
-import './sharetemplates.js'
-import './shareitemmodel.js'
-import './sharesocialmanager.js'
-import './sharedialogresharerinfoview.js'
-import './sharedialoglinkshareview.js'
-import './sharedialogshareelistview.js'
-import './sharedialogview.js'
-import './share.js'
