--- core/js/share/sharedialoglinkshareview.handlebars
@@ -1,40 +1,0 @@
-{{#if shareAllowed}}
-<ul class="shareWithList">
-	{{#if nolinkShares}}
-		<li data-share-id="{{newShareId}}">
-			<div class="avatar icon-public-white"></div>
-			<span class="username">{{newShareLabel}}</span>
-			<span class="sharingOptionsGroup">
-				<div class="share-menu">
-					<a href="#" class="icon icon-add new-share has-tooltip {{#if showPending}}hidden{{/if}}" title="{{newShareTitle}}"></a>
-					<span class="icon icon-loading-small {{#unless showPending}}hidden{{/unless}}"></span>
-					{{#if showPending}}
-						{{{pendingPopoverMenu}}}
-					{{/if}}
-				</div>
-			</span>
-		</li>
-	{{/if}}
-	{{#each linkShares}}
-		<li data-share-id="{{cid}}">
-			<div class="avatar icon-public-white"></div>
-			<span class="username" title="{{linkShareCreationDate}}">{{linkShareLabel}}</span>
-			
-			<span class="sharingOptionsGroup">
-				<a href="#" class="clipboard-button icon icon-clippy has-tooltip" data-clipboard-text="{{shareLinkURL}}" title="{{copyLabel}}"></a>
-				<div class="share-menu">
-					<a href="#" class="icon icon-more {{#if showPending}}hidden{{/if}}"></a>
-					<span class="icon icon-loading-small {{#unless showPending}}hidden{{/unless}}"></span>
-					{{#if showPending}}
-						{{{pendingPopoverMenu}}}
-					{{else}}
-						{{{popoverMenu}}}
-					{{/if}}
-				</div>
-			</span>
-		</li>
-	{{/each}}
-</ul>
-{{else}}
-{{#if noSharingPlaceholder}}<input id="shareWith-{{cid}}" class="shareWithField" type="text" placeholder="{{noSharingPlaceholder}}" disabled="disabled" />{{/if}}
-{{/if}}
