--- core/js/share/sharedialoglinkshareview_popover_menu.handlebars
@@ -1,121 +1,0 @@
-<div class="popovermenu menu">
-	<ul>
-		<li class="hidden linkTextMenu">
-			<span class="menuitem icon-link-text">
-				<input id="linkText-{{cid}}" class="linkText" type="text" readonly="readonly" value="{{shareLinkURL}}" />
-			</span>
-		</li>
-		{{#if publicUpload}}
-			<li>
-				<span class="menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="radio" name="publicUpload" value="{{publicUploadRValue}}" id="sharingDialogAllowPublicUpload-r-{{cid}}" class="radio publicUploadRadio" {{{publicUploadRChecked}}} />
-					<label for="sharingDialogAllowPublicUpload-r-{{cid}}">{{publicUploadRLabel}}</label>
-				</span>
-			</li>
-			<li>
-				<span class="menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="radio" name="publicUpload" value="{{publicUploadRWValue}}" id="sharingDialogAllowPublicUpload-rw-{{cid}}" class="radio publicUploadRadio" {{{publicUploadRWChecked}}} />
-					<label for="sharingDialogAllowPublicUpload-rw-{{cid}}">{{publicUploadRWLabel}}</label>
-				</span>
-			</li>
-			<li>
-				<span class="menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="radio" name="publicUpload" value="{{publicUploadWValue}}" id="sharingDialogAllowPublicUpload-w-{{cid}}" class="radio publicUploadRadio" {{{publicUploadWChecked}}} />
-					<label for="sharingDialogAllowPublicUpload-w-{{cid}}">{{publicUploadWLabel}}</label>
-				</span>
-			</li>
-		{{/if}}
-		{{#if publicEditing}}
-			<li id="allowPublicEditingWrapper">
-				<span class="menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="checkbox" name="allowPublicEditing" id="sharingDialogAllowPublicEditing-{{cid}}" class="checkbox publicEditingCheckbox" {{{publicEditingChecked}}} />
-					<label for="sharingDialogAllowPublicEditing-{{cid}}">{{publicEditingLabel}}</label>
-				</span>
-			</li>
-		{{/if}}
-			<li>
-				<span class="menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="checkbox" name="hideDownload" id="sharingDialogHideDownload-{{cid}}" class="checkbox hideDownloadCheckbox"
-					{{#if hideDownload}}checked="checked"{{/if}} />
-					<label for="sharingDialogHideDownload-{{cid}}">{{hideDownloadLabel}}</label>
-				</span>
-			</li>
-			<li>
-				<span class="menuitem">
-					<input type="checkbox" name="showPassword" id="showPassword-{{cid}}" class="checkbox showPasswordCheckbox"
-					{{#if isPasswordSet}}checked="checked"{{/if}} {{#if isPasswordEnforced}}disabled="disabled"{{/if}} value="1" />
-					<label for="showPassword-{{cid}}">{{enablePasswordLabel}}</label>
-				</span>
-			</li>
-			<li class="{{#unless isPasswordSet}}hidden{{/unless}} linkPassMenu">
-				<span class="menuitem icon-share-pass">
-					<input id="linkPassText-{{cid}}" class="linkPassText" type="password" placeholder="{{passwordPlaceholder}}" autocomplete="new-password" />
-					<input type="submit" class="icon-confirm share-pass-submit" value="" />
-					<span class="icon icon-loading-small hidden"></span>
-				</span>
-			</li>
-		{{#if showPasswordByTalkCheckBox}}
-			<li>
-				<span class="shareOption menuitem">
-					<span class="icon-loading-small hidden"></span>
-					<input type="checkbox" name="passwordByTalk" id="passwordByTalk-{{cid}}" class="checkbox passwordByTalkCheckbox"
-					{{#if isPasswordByTalkSet}}checked="checked"{{/if}} />
-					<label for="passwordByTalk-{{cid}}">{{passwordByTalkLabel}}</label>
-				</span>
-			</li>
-		{{/if}}
-		<li>
-			<span class="menuitem">
-				<input id="expireDate-{{cid}}" type="checkbox" name="expirationDate" class="expireDate checkbox"
-				{{#if hasExpireDate}}checked="checked"{{/if}} {{#if isExpirationEnforced}}disabled="disabled"{{/if}} />
-				<label for="expireDate-{{cid}}">{{expireDateLabel}}</label>
-			</span>
-		</li>
-		<li class="{{#unless hasExpireDate}}hidden{{/unless}}">
-			<span class="menuitem icon-expiredate expirationDateContainer-{{cid}}">
-				<label for="expirationDatePicker-{{cid}}" class="hidden-visually" value="{{expirationDate}}">{{expirationLabel}}</label>
-				<!-- do not use the datepicker if enforced -->
-				<input id="expirationDatePicker-{{cid}}" class="{{#unless isExpirationEnforced}}datepicker{{/unless}}" type="text"
-					placeholder="{{expirationDatePlaceholder}}" value="{{#if hasExpireDate}}{{expireDate}}{{else}}{{defaultExpireDate}}{{/if}}"
-					data-max-date="{{maxDate}}" {{#if isExpirationEnforced}}readonly{{/if}} />
-			</span>
-			</li>
-		<li>
-			<a href="#" class="share-add">
-				<span class="icon-loading-small hidden"></span>
-				<span class="icon icon-edit"></span>
-				<span>{{addNoteLabel}}</span>
-				<input type="button" class="share-note-delete icon-delete {{#unless hasNote}}hidden{{/unless}}">
-			</a>
-		</li>
-		<li class="share-note-form share-note-link {{#unless hasNote}}hidden{{/unless}}">
-			<span class="menuitem icon-note">
-				<textarea class="share-note">{{shareNote}}</textarea>
-				<input type="submit" class="icon-confirm share-note-submit" value="" id="add-note-{{shareId}}" />
-			</span>
-		</li>
-		{{#each social}}
-			<li>
-				<a href="#" class="menuitem pop-up" data-url="{{url}}" data-window="{{newWindow}}">
-					<span class="icon {{iconClass}}"></span>
-					<span>{{label}}</span>
-				</a>
-			</li>
-		{{/each}}
-		<li>
-			<a href="#" class="unshare"><span class="icon-loading-small hidden"></span><span class="icon icon-delete"></span><span>{{unshareLinkLabel}}</span></a>
-		</li>
-		<li>
-			<a href="#" class="new-share">
-				<span class="icon-loading-small hidden"></span>
-				<span class="icon icon-add"></span>
-				<span>{{newShareLabel}}</span>
-			</a>
-		</li>
-	</ul>
-</div>
