--- core/js/share/sharedialoglinkshareview_popover_menu_pending.handlebars
@@ -1,20 +1,0 @@
-<div class="popovermenu open menu pending">
-	<ul>
-		{{#if isPasswordEnforced}}
-			<li>
-				<span class="menuitem icon-info">
-					<p>{{enforcedPasswordLabel}}</p>
-				</span>
-			</li>
-			<li class="linkPassMenu">
-				<span class="menuitem">
-					<form autocomplete="off" class="enforcedPassForm">
-						<input id="enforcedPassText" required class="enforcedPassText" type="password"
-							placeholder="{{passwordPlaceholder}}" autocomplete="enforcedPassText" minlength="{{minPasswordLength}}" />
-						<input type="submit" value=" " class="primary icon-checkmark-white">
-					</form>
-				</span>
-			</li>
-		{{/if}}
-	</ul>
-</div>
