--- core/js/share/sharedialogresharerinfoview.handlebars
@@ -1,5 +1,0 @@
-<span class="reshare">
-	<div class="avatar" data-userName="{{reshareOwner}}"></div>
-	{{sharedByText}}
-</span>
-{{#if hasShareNote}}<div class="share-note">{{shareNote}}</div>{{/if}}
