--- core/js/share/sharedialogshareelistview.handlebars
@@ -1,32 +1,0 @@
-<ul id="shareWithList" class="shareWithList">
-	{{#each sharees}}
-        {{#unless isShareWithCurrentUser}}
-		<li data-share-id="{{shareId}}" data-share-type="{{shareType}}" data-share-with="{{shareWith}}">
-			<div class="avatar {{#if modSeed}}imageplaceholderseed{{/if}}" data-username="{{shareWith}}" data-avatar="{{shareWithAvatar}}" data-displayname="{{shareWithDisplayName}}" {{#if modSeed}}data-seed="{{shareWith}} {{shareType}}"{{/if}}></div>
-			<span class="username" title="{{shareWithTitle}}">{{shareWithDisplayName}}</span>
-			{{#if canUpdateShareSettings }}
-			<span class="sharingOptionsGroup">
-				{{#if editPermissionPossible}}
-					<span>
-						<input id="canEdit-{{cid}}-{{shareId}}" type="checkbox" name="edit" class="permissions checkbox" />
-						<label for="canEdit-{{cid}}-{{shareId}}">{{canEditLabel}}</label>
-					</span>
-				{{/if}}
-				<div tabindex="0" class="share-menu"><span class="icon icon-more"></span>
-					{{{popoverMenu}}}
-				</div>
-			</span>
-			{{/if}}
-		</li>
-	{{/unless}}
-	{{/each}}
-	{{#each linkReshares}}
-		<li data-share-id="{{shareId}}" data-share-type="{{shareType}}">
-			<div class="avatar" data-username="{{shareInitiator}}"></div>
-			<span class="has-tooltip username" title="{{shareInitiator}}">{{shareInitiatorText}}</span>
-			<span class="sharingOptionsGroup">
-				<a href="#" class="unshare"><span class="icon-loading-small hidden"></span><span class="icon icon-delete"></span><span class="hidden-visually">{{unshareLabel}}</span></a>
-			</span>
-		</li>
-	{{/each}}
-</ul>
