--- core/js/share/sharedialogshareelistview_popover_menu.handlebars
@@ -1,107 +1,0 @@
-<div class="popovermenu bubble hidden menu">
-	<ul>
-		{{#if isResharingAllowed}} {{#if sharePermissionPossible}} {{#unless isMailShare}}
-			<li>
-				<span class="menuitem">
-					<input id="canShare-{{cid}}-{{shareId}}" type="checkbox" name="share" class="permissions checkbox" {{#if hasSharePermission}}checked="checked"{{/if}} data-permissions="{{sharePermission}}" />
-					<label for="canShare-{{cid}}-{{shareId}}">{{canShareLabel}}</label>
-				</span>
-				</li>
-			{{/unless}} {{/if}} {{/if}}
-		{{#if isFolder}}
-			{{#if createPermissionPossible}}{{#unless isMailShare}}
-				<li>
-					<span class="menuitem">
-						<input id="canCreate-{{cid}}-{{shareId}}" type="checkbox" name="create" class="permissions checkbox" {{#if hasCreatePermission}}checked="checked"{{/if}} data-permissions="{{createPermission}}"/>
-						<label for="canCreate-{{cid}}-{{shareId}}">{{createPermissionLabel}}</label>
-					</span>
-				</li>
-			{{/unless}}{{/if}}
-			{{#if updatePermissionPossible}}{{#unless isMailShare}}
-				<li>
-					<span class="menuitem">
-						<input id="canUpdate-{{cid}}-{{shareId}}" type="checkbox" name="update" class="permissions checkbox" {{#if hasUpdatePermission}}checked="checked"{{/if}} data-permissions="{{updatePermission}}"/>
-						<label for="canUpdate-{{cid}}-{{shareId}}">{{updatePermissionLabel}}</label>
-					</span>
-				</li>
-				{{/unless}}{{/if}}
-			{{#if deletePermissionPossible}}{{#unless isMailShare}}
-				<li>
-					<span class="menuitem">
-						<input id="canDelete-{{cid}}-{{shareId}}" type="checkbox" name="delete" class="permissions checkbox" {{#if hasDeletePermission}}checked="checked"{{/if}} data-permissions="{{deletePermission}}"/>
-						<label for="canDelete-{{cid}}-{{shareId}}">{{deletePermissionLabel}}</label>
-					</span>
-				</li>
-				{{/unless}}{{/if}}
-			{{/if}}
-		{{#if isMailShare}}
-			{{#if hasCreatePermission}}
-				<li>
-					<span class="menuitem">
-						<input id="secureDrop-{{cid}}-{{shareId}}" type="checkbox" name="secureDrop" class="checkbox secureDrop" {{#if secureDropMode}}checked="checked"{{/if}} data-permissions="{{readPermission}}"/>
-						<label for="secureDrop-{{cid}}-{{shareId}}">{{secureDropLabel}}</label>
-					</span>
-				</li>
-			{{/if}}
-			<li>
-				<span class="menuitem">
-					<input id="password-{{cid}}-{{shareId}}" type="checkbox" name="password" class="password checkbox" {{#if isPasswordSet}}checked="checked"{{/if}}{{#if isPasswordSet}}{{#if isPasswordForMailSharesRequired}}disabled=""{{/if}}{{/if}}" />
-					<label for="password-{{cid}}-{{shareId}}">{{passwordLabel}}</label>
-				</span>
-			</li>
-			<li class="passwordMenu-{{cid}}-{{shareId}} {{#unless isPasswordSet}}hidden{{/unless}}">
-				<span class="passwordContainer-{{cid}}-{{shareId}} icon-passwordmail menuitem">
-					<label for="passwordField-{{cid}}-{{shareId}}" class="hidden-visually" value="{{password}}">{{passwordLabel}}</label>
-					<input id="passwordField-{{cid}}-{{shareId}}" class="passwordField" type="password" placeholder="{{passwordPlaceholder}}" value="{{passwordValue}}" autocomplete="new-password" />
-					<span class="icon-loading-small hidden"></span>
-				</span>
-			</li>
-			{{#if isTalkEnabled}}
-				<li>
-					<span class="menuitem">
-						<input id="passwordByTalk-{{cid}}-{{shareId}}" type="checkbox" name="passwordByTalk" class="passwordByTalk checkbox" {{#if isPasswordByTalkSet}}checked="checked"{{/if}} />
-						<label for="passwordByTalk-{{cid}}-{{shareId}}">{{passwordByTalkLabel}}</label>
-					</span>
-				</li>
-				<li class="passwordByTalkMenu-{{cid}}-{{shareId}} {{#unless isPasswordByTalkSet}}hidden{{/unless}}">
-					<span class="passwordByTalkContainer-{{cid}}-{{shareId}} icon-passwordtalk menuitem">
-						<label for="passwordByTalkField-{{cid}}-{{shareId}}" class="hidden-visually" value="{{password}}">{{passwordByTalkLabel}}</label>
-						<input id="passwordByTalkField-{{cid}}-{{shareId}}" class="passwordField" type="password" placeholder="{{passwordByTalkPlaceholder}}" value="{{passwordValue}}" autocomplete="new-password" />
-						<span class="icon-loading-small hidden"></span>
-					</span>
-				</li>
-				{{/if}}
-			{{/if}}
-		<li>
-			<span class="menuitem">
-				<input id="expireDate-{{cid}}-{{shareId}}" type="checkbox" name="expirationDate" class="expireDate checkbox" {{#if hasExpireDate}}checked="checked"{{/if}}" />
-				<label for="expireDate-{{cid}}-{{shareId}}">{{expireDateLabel}}</label>
-			</span>
-		</li>
-		<li class="expirationDateMenu-{{cid}}-{{shareId}} {{#unless hasExpireDate}}hidden{{/unless}}">
-			<span class="expirationDateContainer-{{cid}}-{{shareId}} icon-expiredate menuitem">
-				<label for="expirationDatePicker-{{cid}}-{{shareId}}" class="hidden-visually" value="{{expirationDate}}">{{expirationLabel}}</label>
-				<input id="expirationDatePicker-{{cid}}-{{shareId}}" class="datepicker" type="text" placeholder="{{expirationDatePlaceholder}}" value="{{#if hasExpireDate}}{{expireDate}}{{else}}{{defaultExpireDate}}{{/if}}" />
-			</span>
-		</li>
-		{{#if isNoteAvailable}}
-			<li>
-				<a href="#" class="share-add">
-					<span class="icon-loading-small hidden"></span>
-					<span class="icon icon-edit"></span>
-					<span>{{addNoteLabel}}</span>
-					<input type="button" class="share-note-delete icon-delete {{#unless hasNote}}hidden{{/unless}}">
-				</a>
-			</li>
-			<li class="share-note-form {{#unless hasNote}}hidden{{/unless}}">
-				<span class="menuitem icon-note">
-					<textarea class="share-note">{{shareNote}}</textarea>
-					<input type="submit" class="icon-confirm share-note-submit" value="" id="add-note-{{shareId}}" />
-				</span>
-			</li>
-		{{/if}}
-		<li>
-			<a href="#" class="unshare"><span class="icon-loading-small hidden"></span><span class="icon icon-delete"></span><span>{{unshareLabel}}</span></a>
-		</li>
-	</ul>
-</div>
