--- core/js/share/sharedialogview.handlebars
@@ -1,12 +1,0 @@
-<div class="resharerInfoView subView"></div>
-{{#if isSharingAllowed}}
-	<label for="shareWith-{{cid}}" class="hidden-visually">{{shareLabel}}</label>
-	<div class="oneline">
-		<input id="shareWith-{{cid}}" class="shareWithField" type="text" placeholder="{{sharePlaceholder}}" />
-		<span class="shareWithLoading icon-loading-small hidden"></span>
-		<span class="shareWithConfirm icon icon-confirm"></span>
-	</div>
-{{/if}}
-<div class="linkShareView subView"></div>
-<div class="shareeListView subView"></div>
-<div class="loading hidden" style="height: 50px"></div>
